tidy up logger
